excluir promotor

+++ API/app/Http/Controllers/PromotoresController.php
@@ -70,8 +70,16 @@ class PromotoresController extends Controller
     public function destroy($id)
     {
         $promotor = Promotores::find($id);
+        if ($promotor) {
+        $userId = $promotor->user_id;
+        $user = User::find($userId);
         $promotor->delete();
+        $user->delete();
         return response()->json(null, 204);
+        } else {
+            // O promotor não foi encontrado, trate o erro
+            abort(404);
+            }
     }
 
     public function update(Request $request, $id)
